AtomicInt: wrap sync/atomic.Int64 w/ go 1.19
Leverage sync/atomic.Int64 with go 1.19 (and keep the same code for
earlier versions)

Use build tags as appropriate.
