+++ atomic_int_go1.19.go
@@ -1,0 +1,44 @@
+//go:build go1.19
+
+/*
+Copyright 2022 Vimeo Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package galaxycache
+
+import (
+	"strconv"
+	"sync/atomic"
+)
+
+// An AtomicInt is an int64 to be accessed atomically.
+// It thinly wraps atomic.Int64 with go1.19+
+type AtomicInt struct {
+	v atomic.Int64
+}
+
+// Add atomically adds n to i.
+func (i *AtomicInt) Add(n int64) {
+	i.v.Add(n)
+}
+
+// Get atomically gets the value of i.
+func (i *AtomicInt) Get() int64 {
+	return i.v.Load()
+}
+
+func (i *AtomicInt) String() string {
+	return strconv.FormatInt(i.Get(), 10)
+}
